requirements.txt
flask
pillow
gunicorn
requests
flask-cors
pytz
pdf2image 
PyMuPDF